feat: add runner abstraction layer for multi-framework agent support
Introduce a pluggable runner architecture so the chat service can
dispatch to different agent frameworks (LlamaStack, LangGraph, CrewAI)
based on the agent's runner_type.

- Add runner_type column to virtual_agents table (default: "llamastack")
- Add runner_type to Pydantic schemas (VirtualAgentBase, VirtualAgentUpdate)
- Create BaseRunner abstract class defining the streaming interface
- Extract all LlamaStack logic into LlamaStackRunner
- Rewrite ChatService as a thin dispatcher that delegates to runners
- Add Alembic migration for the new column
- Add 24 unit tests covering dispatcher, base class, model, and schemas
- Fix dev container build: replace dnf nmap-ncat with Python socket check
