+++ backend/app/services/runners/__init__.py
@@ -1,0 +1,15 @@
+"""
+Runner implementations for different agent frameworks.
+
+Each runner implements the BaseRunner interface to provide a unified
+streaming contract for the chat service, regardless of the underlying
+agent framework (LlamaStack, LangGraph, CrewAI, etc.).
+"""
+
+from .base import BaseRunner
+from .llamastack_runner import LlamaStackRunner
+
+__all__ = [
+    "BaseRunner",
+    "LlamaStackRunner",
+]
